stock_bot/requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fastapi==0.115.11
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
magic-filter==1.0.12
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.2.0
multitasking==0.0.11
namex==0.0.8
narwhals==1.31.0
numpy==2.2.3
openai==1.66.3
optree==0.14.1
packaging==24.2
pandas==2.2.3
patsy==1.0.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
statsmodels==0.14.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
xgboost==3.0.0
yarl==1.18.3
yfinance==0.2.54
python-dotenv==1.0.1
